image_conversion/requirements.txt
pillow
pillow-heif